+++ common/feature-flags.ts
@@ -3,7 +3,7 @@ type FeatureFlags = {
   SHOW_UPSCAYL_CLOUD_INFO: boolean;
 };
 
-export const featureFlags: FeatureFlags = {
+export const FEATURE_FLAGS: FeatureFlags = {
   APP_STORE_BUILD: false,
   SHOW_UPSCAYL_CLOUD_INFO: false,
 };
