+++ electron/index.ts
@@ -2,7 +2,7 @@ import prepareNext from "electron-next";
 import { autoUpdater } from "electron-updater";
 import log from "electron-log";
 import { app, ipcMain, protocol } from "electron";
-import COMMAND from "../common/commands";
+import ELECTRON_COMMANDS from "../common/commands";
 import logit from "./utils/logit";
 import openFolder from "./commands/open-folder";
 import stop from "./commands/stop";
@@ -17,7 +17,7 @@ import { execPath, modelsPath } from "./utils/get-resource-paths";
 import batchUpscayl from "./commands/batch-upscayl";
 import doubleUpscayl from "./commands/double-upscayl";
 import autoUpdate from "./commands/auto-update";
-import { featureFlags } from "../common/feature-flags";
+import { FEATURE_FLAGS } from "../common/feature-flags";
 import settings from "electron-settings";
 
 // INITIALIZATION
@@ -43,14 +43,14 @@ app.on("ready", async () => {
   log.info(
     "🆙 Upscayl version:",
     app.getVersion(),
-    featureFlags.APP_STORE_BUILD && "MAC-APP-STORE",
+    FEATURE_FLAGS.APP_STORE_BUILD && "MAC-APP-STORE",
   );
   log.info("🚀 UPSCAYL EXEC PATH: ", execPath);
   log.info("🚀 MODELS PATH: ", modelsPath);
 
   let closeAccess;
   const folderBookmarks = await settings.get("folder-bookmarks");
-  if (featureFlags.APP_STORE_BUILD && folderBookmarks) {
+  if (FEATURE_FLAGS.APP_STORE_BUILD && folderBookmarks) {
     logit("🚨 Folder Bookmarks: ", folderBookmarks);
     try {
       closeAccess = app.startAccessingSecurityScopedResource(
@@ -68,29 +68,32 @@ app.on("window-all-closed", () => {
 });
 
 // ! ENABLE THIS FOR MACOS APP STORE BUILD
-if (featureFlags.APP_STORE_BUILD) {
+if (FEATURE_FLAGS.APP_STORE_BUILD) {
   logit("🚀 APP STORE BUILD ENABLED");
   app.commandLine.appendSwitch("in-process-gpu");
 }
 
-ipcMain.on(COMMAND.STOP, stop);
+ipcMain.on(ELECTRON_COMMANDS.STOP, stop);
 
-ipcMain.on(COMMAND.OPEN_FOLDER, openFolder);
+ipcMain.on(ELECTRON_COMMANDS.OPEN_FOLDER, openFolder);
 
-ipcMain.handle(COMMAND.SELECT_FOLDER, selectFolder);
+ipcMain.handle(ELECTRON_COMMANDS.SELECT_FOLDER, selectFolder);
 
-ipcMain.handle(COMMAND.SELECT_FILE, selectFile);
+ipcMain.handle(ELECTRON_COMMANDS.SELECT_FILE, selectFile);
 
-ipcMain.on(COMMAND.GET_MODELS_LIST, getModelsList);
+ipcMain.on(ELECTRON_COMMANDS.GET_MODELS_LIST, getModelsList);
 
-ipcMain.handle(COMMAND.SELECT_CUSTOM_MODEL_FOLDER, customModelsSelect);
+ipcMain.handle(
+  ELECTRON_COMMANDS.SELECT_CUSTOM_MODEL_FOLDER,
+  customModelsSelect,
+);
 
-ipcMain.on(COMMAND.UPSCAYL, imageUpscayl);
+ipcMain.on(ELECTRON_COMMANDS.UPSCAYL, imageUpscayl);
 
-ipcMain.on(COMMAND.FOLDER_UPSCAYL, batchUpscayl);
+ipcMain.on(ELECTRON_COMMANDS.FOLDER_UPSCAYL, batchUpscayl);
 
-ipcMain.on(COMMAND.DOUBLE_UPSCAYL, doubleUpscayl);
+ipcMain.on(ELECTRON_COMMANDS.DOUBLE_UPSCAYL, doubleUpscayl);
 
-if (!featureFlags.APP_STORE_BUILD) {
+if (!FEATURE_FLAGS.APP_STORE_BUILD) {
   autoUpdater.on("update-downloaded", autoUpdate);
 }
