+++ electron/utils/logit.ts
@@ -1,5 +1,5 @@
 import log from "electron-log";
-import COMMAND from "../../common/commands";
+import ELECTRON_COMMANDS from "../../common/commands";
 import { getMainWindow } from "../main-window";
 
 const logit = (...args: any) => {
@@ -9,7 +9,7 @@ const logit = (...args: any) => {
   if (process.env.NODE_ENV === "development") {
     return;
   }
-  mainWindow.webContents.send(COMMAND.LOG, args.join(" "));
+  mainWindow.webContents.send(ELECTRON_COMMANDS.LOG, args.join(" "));
 };
 
 export default logit;
