+++ renderer/components/hooks/use-upscayl-version.ts
@@ -1,0 +1,16 @@
+import { useState, useEffect } from "react";
+
+const useUpscaylVersion = () => {
+  const [version, setVersion] = useState<string | null>(null);
+
+  useEffect(() => {
+    const upscaylVersion = navigator?.userAgent?.match(
+      /Upscayl\/([\d\.]+\d+)/,
+    )?.[1];
+    setVersion(upscaylVersion);
+  }, []);
+
+  return version;
+};
+
+export default useUpscaylVersion;
