+++ renderer/lib/valid-formats.ts
@@ -1,0 +1,1 @@
+export const VALID_IMAGE_FORMATS = ["png", "jpg", "jpeg", "jfif", "webp"];
