+++ renderer/renderer.d.ts
@@ -2,6 +2,7 @@ import { IpcRenderer } from "electron";
 
 export interface IElectronAPI {
   on: (command, func?) => IpcRenderer;
+  off: (command, func?) => IpcRenderer;
   send: <T>(command, func?: T) => IpcRenderer;
   invoke: (command, func?) => any;
   platform: "mac" | "win" | "linux";
